LogExpFunctionsExt.jl Documentation

LogExpFunctionsExt.jl Documentation

Types

Functions

Index

Types

Functions

LogExpFunctionsExt.log1mlogistic — Method

log1mlogistic(x)

The natural logarithm of the 1 minus the inverse logit function, computed more carefully and with fewer function calls than log(1 - logistic(x)).

Its inverse is the logit1mexp function.

LogExpFunctionsExt.logit1mexp — Method

logit1mexp(x)

The logit of 1 minus the exponential of x, computed more carefully and with fewer function calls than logit(1 - exp(x)).

Its inverse is the log1mlogistic function.

LogExpFunctionsExt.logitexp — Method

logitexp(x)

The logit of the exponential of x, computed more carefully and with fewer function calls than logit(exp(x))

Its inverse is the loglogistic function.

LogExpFunctionsExt.loglogistic — Method

loglogistic(x)

The natural logarithm of the logistic function, computed more carefully and with fewer calls than than the composition log(logistic(x)).

Its inverse is the logitexp function.

Index

LogExpFunctionsExt.log1mlogistic

LogExpFunctionsExt.logit1mexp

LogExpFunctionsExt.logitexp

LogExpFunctionsExt.loglogistic